Fix pi0 and pi0fast infer issue
  1. Fix loss_dict format
  2. Passin the prompt text for pi0 policy
  3. Correct data type for pi0fast

+++ lerobot/common/policies/pi0fast/modeling_pi0fast.py
@@ -798,7 +798,7 @@ class PI0FAST(nn.Module):
                 print(f"Error decoding tokens: {e}")
                 print(f"Tokens: {token}")
                 decoded_dct_coeff = np.zeros((self.time_horizon, self.action_dim))
-            decoded_actions.append(idct(decoded_dct_coeff / self.fast_tokenizer.scale, axis=0, norm="ortho"))
+            decoded_actions.append(idct(decoded_dct_coeff / self.fast_tokenizer.scale, axis=0, norm="ortho").astype(np.float32))
         return np.stack(decoded_actions)
 
     def extract_actions(self, tokens: torch.Tensor, action_horizon: int, action_dim: int) -> torch.Tensor:
